requirements.txt
python-dotenv
inquirer